[Refactor] directory-structure follow by directory-structure.md
