+++ js/game.js
@@ -1,33 +1,36 @@
-import { DROPSPEED, currentTetrimino, addNewTetrimino } from './tetrimino.js';
-import { drawPlayScreen, drawTetrimino } from './renderer.js';
+import { PLAY_SCREEN_WIDTH, PLAY_SCREEN_HEIGHT, DROP_SPEED } from './utils.js';
+import { initField, field, clearFullLines } from './board.js';
+import { currentTetrimino, addNewTetrimino, moveTetrimino, lockTetrimino } from './tetrimino.js';
+import { drawPlayScreen } from './renderer.js';
+import { checkGameOver, handleGameOver } from './score.js';
 
+export let animationId;
 let lastDropTime = 0;
 
 export function initGame() {
-    // ゲームの初期化処理
+    initField();
 }
 
 export function gameLoop(currentTime) {
-    // ゲームループ関数
-    // requestAnimationFlameの呼び出しにより現在の時刻:DOMHighResTimeStampがgameLoopに渡される
-    // それをcurrentTimeとして受け取っている
     drawPlayScreen();
     if (currentTetrimino === null) {
         addNewTetrimino();
+        if (checkGameOver()) {
+            handleGameOver();
+            return;
+        }
         lastDropTime = currentTime;
     } else {
         normalDrop(currentTime);
-        drawTetrimino();
     }
-
-    requestAnimationFrame(gameLoop);
+    animationId = requestAnimationFrame(gameLoop);
 }
 
 function normalDrop(currentTime) {
-    // ...(テトリミノを1行落下させる関数)
-    // 現在時刻が最終ドロップ時刻よりもDROPSPEED経過していれば落下
-    if (currentTime - lastDropTime > DROPSPEED) {
-        currentTetrimino.row++;
+    if (currentTime - lastDropTime > DROP_SPEED) {
+        if (!moveTetrimino(currentTetrimino.row + 1, currentTetrimino.column)) {
+            lockTetrimino();
+        }
         lastDropTime = currentTime;
     }
 }
