+++ js/input.js
@@ -1,22 +1,18 @@
-import {currentTetrimino, shiftLeft, shiftRight, rotateTetrimino } from './tetrimino.js';
+import { moveLeft, moveRight, moveDown, hardDrop, rotateTetrimino } from './tetrimino.js';
 
 export function initInput() {
     document.addEventListener('keydown', handleKeyDown);
 }
 
 function handleKeyDown(e) {
-    if (!currentTetrimino) return;
-
     switch(e.key) {
-
-        case "ArrowLeft": //"ArrowLeft"は左矢印キーの識別子になります。
-            shiftLeft();
-            break;
-        case "ArrowRight": //"ArrowRight"は左矢印キーの識別子になります。
-            shiftRight();
-            break;
-        case "ArrowUp": // 追加: 上矢印キーで回転
-            rotateTetrimino();
-            break;
+        case "ArrowLeft": moveLeft(); break;
+        case "ArrowRight": moveRight(); break;
+        case "ArrowUp":
+        case "x": rotateTetrimino(true); break;
+        case "Control":
+        case "z": rotateTetrimino(false); break;
+        case "ArrowDown": moveDown(); break;
+        case " ": hardDrop(); break;
     }
 }
