+++ js/tetrimino.js
@@ -1,159 +1,123 @@
-import { PLAYSCREENWIDTH } from './utils.js';
+import { TETRIMINOS, PLAY_SCREEN_HEIGHT, PLAY_SCREEN_WIDTH } from './utils.js';
+import { field, checkCollision, clearFullLines } from './board.js';
 
-// テトリミノの通常落下速度(millisec)
-export const DROPSPEED = 1000;
-
-// 現在のテトリミノを保持する変数
 export let currentTetrimino = null;
-
-const TETROMINOS = {
-    // ... (テトリミノの定義)
-    I: {
-        shape: [
-            [0, 0, 0, 0],
-            [1, 1, 1, 1],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        color: '#00FFFF'
-    },
-    O: {
-        shape: [
-            [1, 1],
-            [1, 1]
-        ],
-        color: '#FFFF00'
-    },
-    T: {
-        shape: [
-            [0, 1, 0],
-            [1, 1, 1],
-            [0, 0, 0]
-        ],
-        color: '#800080'
-    },
-    S: {
-        shape: [
-            [0, 1, 1],
-            [1, 1, 0],
-            [0, 0, 0]
-        ],
-        color: '#00FF00'
-    },
-    Z: {
-        shape: [
-            [1, 1, 0],
-            [0, 1, 1],
-            [0, 0, 0]
-        ],
-        color: '#FF0000'
-    },
-    J: {
-        shape: [
-            [1, 0, 0],
-            [1, 1, 1],
-            [0, 0, 0]
-        ],
-        color: '#0000FF'
-    },
-    L: {
-        shape: [
-            [0, 0, 1],
-            [1, 1, 1],
-            [0, 0, 0]
-        ],
-        color: '#FFA500'
-    }
-};
-
-
-// テトリミノの生成順序
 let tetriminoSequence = [];
 
-// 次のテトリミノを取得する関数
 export function getNextTetrimino() {
     if (tetriminoSequence.length === 0) {
-        // 新しい7種のテトリミノをランダム順に生成
-        tetriminoSequence = ['I', 'O', 'T', 'S', 'Z', 'J', 'L'].sort(() => Math.random() - 0.5);
+        tetriminoSequence = Object.keys(TETRIMINOS).sort(() => Math.random() - 0.5);
     }
     const name = tetriminoSequence.pop();
-    const matrix = TETROMINOS[name].shape;
-    const color = TETROMINOS[name].color;
-
-    // 初期位置の設定
-    const column = PLAYSCREENWIDTH / 2 - Math.ceil(matrix[0].length / 2);
+    const { shape, color } = TETRIMINOS[name];
+    const column = Math.floor(PLAY_SCREEN_WIDTH / 2) - Math.ceil(shape[0].length / 2);
     const row = 0;
 
-    return {
-        name: name,     // テトリミノの種類
-        matrix: matrix, // 形状
-        color: color,   // 色
-        row: row,       // 現在の行
-        column: column  // 現在の列
-    };
+    return { name, shape, color, row, column };
 }
 
-// 新しいテトリミノをゲームに追加する関数
 export function addNewTetrimino() {
     currentTetrimino = getNextTetrimino();
-    // drawTetrimino();
 }
 
-// / テトリミノを左に移動させる関数
-export function shiftLeft() {
-    currentTetrimino.column--;
-}
-
-// テトリミノを右に移動させる関数
-export function shiftRight() {
-    currentTetrimino.column++;
-}
-
-export function rotateTetrimino() {
-    const iTetriminoSize = 4;
-    
-    // 現在のテトリミノの形状を取得
-    let currentShape = currentTetrimino.matrix;
-    
-    // 現在の形状の大きさ（matrixSize x matrixSize）を取得
-    let matrixSize = currentShape.length;
-    
-    // 新しい空の配列を作成（回転後の形状を格納するため）
-    let rotatedShape = [];
-    
-    // 新しい形状の配列を0で初期化
-    for (let i = 0; i < matrixSize; i++) {
-        let newRow = [];
-        for (let j = 0; j < matrixSize; j++) {
-            newRow.push(0);
-        }
-        rotatedShape.push(newRow);
+export function moveTetrimino(newRow, newColumn, newShape = null) {
+    const originalTetrimino = { ...currentTetrimino };
+    currentTetrimino.row = newRow;
+    currentTetrimino.column = newColumn;
+    if (newShape) {
+        currentTetrimino.shape = newShape;
     }
-    
-    // I型テトリミノの場合は特別な回転を行う
-    if (currentTetrimino.name === 'I') {
-        // I型テトリミノが横長の場合
-        if (currentShape[1][0] === 1) {
-            // 縦長に変更
-            for (let i = 0; i < iTetriminoSize; i++) {
-                rotatedShape[i][2] = 1;
+
+    if (checkCollision(currentTetrimino)) {
+        Object.assign(currentTetrimino, originalTetrimino);
+        return false;
+    }
+    return true;
+}
+
+// テトリミノの移動操作
+export const moveLeft = () => moveTetrimino(currentTetrimino.row, currentTetrimino.column - 1);
+export const moveRight = () => moveTetrimino(currentTetrimino.row, currentTetrimino.column + 1);
+export const moveDown = () => moveTetrimino(currentTetrimino.row + 1, currentTetrimino.column);
+
+// テトリミノの落下地点を取得
+export function getTetriminoDropPosition() {
+    let ghostRow = currentTetrimino.row;
+    while (!checkCollision({ ...currentTetrimino, row: ghostRow + 1 })) {
+        ghostRow++;
+    }
+    return ghostRow;
+}
+
+// ハードドロップ
+export function hardDrop() {
+    const dropRow = getTetriminoDropPosition();
+    if (moveTetrimino(dropRow, currentTetrimino.column)) {
+        lockTetrimino();
+        currentTetrimino = null;
+    }
+}
+
+export function lockTetrimino() {
+    const { shape, row, column, color } = currentTetrimino;
+    shape.forEach((rowShape, r) => {
+        rowShape.forEach((cell, c) => {
+            if (cell) {
+                const y = row + r;
+                const x = column + c;
+                if (y >= 0 && y < PLAY_SCREEN_HEIGHT && x >= 0 && x < PLAY_SCREEN_WIDTH) {
+                    field[y][x] = color;
+                }
             }
-        } else {
-            // 横長に変更
-            for (let i = 0; i < iTetriminoSize; i++) {
-                rotatedShape[1][i] = 1;
+        });
+    });
+    clearFullLines();
+    currentTetrimino = null;
+}
+
+export function rotateTetrimino(clockwise = true) {
+    const { name, shape } = currentTetrimino;
+    let newShape;
+
+    if (name === 'O') return; // O型は回転しない
+
+    const iTetriminoSize = 4;
+
+    if (name === 'I') {
+        newShape = Array(iTetriminoSize).fill().map(() => Array(iTetriminoSize).fill(0));
+        
+        if (clockwise) {
+            if (shape[1][0] === 1) { // 横長の場合
+                // 縦長に変更
+                for (let i = 0; i < iTetriminoSize; i++) {
+                    newShape[i][2] = 1;
+                }
+            } else { // 縦長の場合
+                // 横長に変更
+                for (let i = 0; i < iTetriminoSize; i++) {
+                    newShape[1][i] = 1;
+                }
+            }
+        } else { // 反時計回り
+            if (shape[0][1] === 1 || shape[0][2] === 1) { // 縦長の場合
+                // 横長に変更
+                for (let i = 0; i < iTetriminoSize; i++) {
+                    newShape[2][i] = 1;
+                }
+            } else { // 横長の場合
+                // 縦長に変更
+                for (let i = 0; i < iTetriminoSize; i++) {
+                    newShape[i][1] = 1;
+                }
             }
         }
     } else {
-        // I型以外のテトリミノの回転
-        for (let i = 0; i < matrixSize; i++) {
-            for (let j = 0; j < matrixSize; j++) {
-                // 90度回転の計算：新しい位置 = [j][matrixSize-1-i]
-                rotatedShape[j][matrixSize - 1 - i] = currentShape[i][j];
-            }
-        }
+        // その他のテトリミノの回転処理
+        newShape = shape.map((row, i) => 
+            row.map((_, j) => clockwise ? shape[shape.length - 1 - j][i] : shape[j][shape.length - 1 - i])
+        );
     }
 
-    // 回転後の形状を現在のテトリミノに設定
-    currentTetrimino.matrix = rotatedShape;
+    moveTetrimino(currentTetrimino.row, currentTetrimino.column, newShape);
+
 }
