+++ js/utils.js
@@ -1,10 +1,66 @@
-// 1ブロックの大きさ
-export const BLOCKSIZE = 30;
-// フィールドのサイズ
-export const PLAYSCREENWIDTH = 10;
-export const PLAYSCREENHEIGHT = 22;
-// キャンバスサイズ（＝プレイ画面のサイズ）
-export const CANVASWIDTH = BLOCKSIZE * PLAYSCREENWIDTH;
-export const CANVASHEIGHT = BLOCKSIZE * PLAYSCREENHEIGHT;
+export const BLOCK_SIZE = 30;
+export const PLAY_SCREEN_WIDTH = 10;
+export const PLAY_SCREEN_HEIGHT = 22;
+export const CANVAS_WIDTH = BLOCK_SIZE * PLAY_SCREEN_WIDTH;
+export const CANVAS_HEIGHT = BLOCK_SIZE * PLAY_SCREEN_HEIGHT;
+export const DROP_SPEED = 1000;
 
-// 他の汎用的な関数や定数をここに定義
+// テトリミノの定義
+export const TETRIMINOS = {
+    I: {
+        shape: [
+            [0, 0, 0, 0],
+            [1, 1, 1, 1],
+            [0, 0, 0, 0],
+            [0, 0, 0, 0]
+        ],
+        color: '#00FFFF'
+    },
+    O: {
+        shape: [
+            [1, 1],
+            [1, 1]
+        ],
+        color: '#FFFF00'
+    },
+    T: {
+        shape: [
+            [0, 1, 0],
+            [1, 1, 1],
+            [0, 0, 0]
+        ],
+        color: '#800080'
+    },
+    S: {
+        shape: [
+            [0, 1, 1],
+            [1, 1, 0],
+            [0, 0, 0]
+        ],
+        color: '#00FF00'
+    },
+    Z: {
+        shape: [
+            [1, 1, 0],
+            [0, 1, 1],
+            [0, 0, 0]
+        ],
+        color: '#FF0000'
+    },
+    J: {
+        shape: [
+            [1, 0, 0],
+            [1, 1, 1],
+            [0, 0, 0]
+        ],
+        color: '#0000FF'
+    },
+    L: {
+        shape: [
+            [0, 0, 1],
+            [1, 1, 1],
+            [0, 0, 0]
+        ],
+        color: '#FFA500'
+    }
+};
